// src/types/movieTypes.ts
export type movieType = {
  adult: boolean;
  backdrop_path: string;
  genre_ids?: number[];
  id: number;
  media_type?: string;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};


export type singleMovieType = {
  adult: false
  backdrop_path: string
  belongs_to_collection: {id: number, name: string, poster_path: string, backdrop_path: string}
  budget: number
  genres: {id: number, name: string}[]
  homepage: string
  id: number
  imdb_id: string
  original_language: string
  original_title: string
  overview: string
  popularity: number
  poster_path: string
  production_companies: {id: number, logo_path: string, name: string, origin_country: string}[]
  production_countries: {}[]
  release_date: string
  revenue: number
  runtime: number
  spoken_languages: {}[]
  status: string
  tagline: string
  title: string
  video: boolean
  vote_average: number
  vote_count: number
}